Change the path for every item of the subdirectories in a directory rename
Summary:
As only the items in subdirectories of the renamed one are modified, if all the items are changed the same part of the path with the same new value, the order is preserved, therefore modify the path directly in the list.

BUG: 401552

Test Plan:
Can't reproduce the bug.
Passes kdirmodeltest and kdirlistertest (and many others).

Reviewers: dfaure, #frameworks

Reviewed By: dfaure

Subscribers: elvisangelaccio, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17619