properly attempt to locate the kioslave bin in $libexec AND $libexec/kf5
Summary:
the kioslave harness is installed to LIBEXEC_KF5 which is '$libexec/' on
win32 (technically $bindir) and '$libexec/kf5/' everywhere else.
since we cannot expect qt.conf to point to the $libexec/kf5/ path
(what with other bits of the application or libraries wanting to find
their own stuff in libexec), we need to explicitly try to search the
suffixed version as well. IOW: the expectation is that the path is
`/foo/libexec/` and we'll also want to check `/foo/libexec/kf5/` as that
is where the binary may be located on !win32.

this enables !win32 builds to be somewhat more relocatable, so long as
qt.conf contains a suitable path config

Reviewers: davidedmundson, dfaure, apol

Reviewed By: apol

Subscribers: apol, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D18725