Fix "Invalid URL: QUrl("some.txt")" warnings in Save dialog
When the Save File dialog is in use with filters, any relative path
would trigger warnings due to KIO::stat receiving an invalid URL in
KFileWidgetPrivate::updateLocationEditExtension. Fix the warning by
ensuring that relative paths are valid file URLs.

Add a bunch of tests to check the expected properties while at it. These
tests pass regardless of the fix but after the fix, warnings are gone.

BUG: 373119

Differential Revision: https://phabricator.kde.org/D19027