Make file overwrite a bit safer
Summary:
Instead of truncating the existing file, a separate copy of the source file in the
destination folder is created. If the copy was successfully created, the existing
destination file is replaced with it. If in middle of creating the separate copy the
disk runs out of space then an attempt is made to delete the existing destination.
After that is done the copy operation is resumed.

BUG: 125102
FIXED-IN: 5.56

Test Plan: Unit testing and manual testing done.

Reviewers: dfaure, ngraham

Reviewed By: dfaure, ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D19080