kjobs: Fix compiler warning for uninit value in SimpleJobPrivate.
GCC 8.2 gives a warning about using an uninitialized value in
SimpleJobPrivate, in a factory function for SimpleJobs that can also
apply relevant JobFlags based on the job's command.

The warning is because the compiler is unable to recognize that the
'command' flag is only able to take the three listed values in this
codepath. Warning fixed by adding a default case.

BUG:404474
FIXED-IN: 5.57

Differential Revision: https://phabricator.kde.org/D20008