[KUrlNavigator] Fix URL navigation when exiting archive with krarc and Dolphin
Summary:
If Krusader is installed, the `krarc` protocol is available and Dolphin may use it to open ZIP archives. In this case, when the user goes back to a parent folder outside the archive, a "file or folder does not exist" error will be returned. This can be fixed by including krarc in the URL scheme check inside the `KUrlNavigator::setLocationUrl` method.

BUG: 386448
FIXED-IN: 5.58

Test Plan:
1. Install Krusader and Dolphin.
2. Enable "Open archives as folder" in Dolphin and open a zip file using the `krarc` protocol. (Change `zip:` to `karc:` when applicable.)
3. Click on the breadcrumb in the location area for a parent folder which is outside the archive.

Reviewers: kde-frameworks-devel, ngraham

Reviewed By: ngraham

Subscribers: bcooksley, aacid, ngraham

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D20424