[FileWidget] Replace "Filter:" with "File type:" when saving with a limited list of mimetypes
Summary:
The Filter combobox currently always has the label "Filter:" This is appropriate in most
cases, but becomes inaccurate when saving with a limited list of mimetypes. In this case,
the combobox loses its text field and becomes a file type chooser rather than a filter.

This patch changes the label to "File type:" for that use case.

BUG: 79903
FIXED-IN: 5.59

Test Plan:
Opening a file: no changes
{F6803789}

Saving a file with no mimetype filter: no changes
{F6803788}

Saving a file with a limited list of mimetypes: label is changed
{F6803790}

Reviewers: #vdg, elvisangelaccio, GB_2

Reviewed By: elvisangelaccio

Subscribers: pino, apol, kde-frameworks-devel

Tags: #frameworks

Maniphest Tasks: T8552

Differential Revision: https://phabricator.kde.org/D20964