Add Documents to the default list of Places
Summary:
This adds Documents to the default list of Places, as discussed with
general favor in D10245. This adds a common XDG directory that is often
used by users, making it easier and faster by default to access files.

Test Plan:
Created a new user and observed and clicked on Documents, and the other
default places locations.

Reviewers: #frameworks, #dolphin, ngraham

Reviewed By: #dolphin, ngraham

Subscribers: davidc, elvisangelaccio, kde-frameworks-devel, ngraham

Tags: #frameworks

Maniphest Tasks: T8349

Differential Revision: https://phabricator.kde.org/D16852