Fix desktop link to file or directory
Summary: KUrlRequester: Show a popup to select between File and Directory before displaying the file dialog.

Test Plan:
Right click onto desktop, select 'Create new' and then
'Basic link to file or directory' or 'Link to Location (URL)'.
Click the file selection icon. A popup is displayed.

If 'File' is selected, a file dialog is displayed, showing files and directories, allowing the selection of a file.
If 'Directory' is selected, a file dialog is displayed, showing only directories, allowing the selection of a directory.

Without the patch, no popup is displayed, only directories are shown in the file dialog and only a directory can be selected.

BUG: 357171
FIXED-IN: 5.59

Reviewers: #frameworks, ngraham, dfaure

Reviewed By: ngraham, dfaure

Subscribers: bruns, michaelweghorn, ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D20197