Show feedback inline when creating new files or folders
Summary:
Right now, the new file/folder dialog allows you to give it an invalid name, and
afterwards it will show you a dialog window complaining at you. This is not very
user-friendly, and becomes more user-unfriendly if we want to use it to show
warnings rather than errors. Finally it doesn't scale well because as we add more
conditions to check for, we'll need to add more dialog windows.

This patch implements the name feedback/warning/error inline in the original name
dialog itself, and disables the {nav OK} button for invalid names.

This patch obsoletes D18384, D18563, and D18599.

Test Plan: {F6915347}

Reviewers: #vdg, #frameworks, shubham, filipf, meven

Reviewed By: #vdg, filipf, meven

Subscribers: filipf, squeakypancakes, dhaumann, aacid, meven, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D21907