KRun: remove KProcessRunner dependency on QtWidgets
Summary:
The messagebox has moved up to the user of the class, KRun,
via a signal.

The long term plan is to move KProcessRunner to KIOGui, so the future
OpenUrlJob can use it, but KProcessRunner uses KWindowSystem which still
depends on QtWidgets at the moment.

Test Plan:
wrote ~/.local/share/applications/appdoesnotexist.desktop
with Exec=appdoesnotexist and MimeType=text/plain;
Started dolphin, RMB on text file, Open With AppDoesNotExist, and the message
box "Could not find the program 'appdoesnotexist'" showed up as expected.

Reviewers: nicolasfella, vkrause

Reviewed By: vkrause

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D23092