Added dialog to set execute permission for executable file when trying to run it.
Summary:
Currently if you try to run executable file for which you don't have execute permission, KRun will either open it in some default program, or prompt user to choose one. It would make sense that, if a file can be normally run with +x permission, user should be asked if he likes to automatically make it executable and run it. If user chooses "Open" option, KRun falls back to default behavior.
This QoL change makes it easier to just download & run things like AppImage files.
Obviously, running executables downloaded from the internet may not be safe, so a warning is included in the prompt.

Here is how it looks:
{F7309791}

Test Plan:
1. Obtain two executable files (e.g. AppImage), set +x on one and -x on another.
2. Try to run both files. The one with +x will run normally, the other one will show the prompt.

Reviewers: #frameworks, dfaure, cfeck, pino

Reviewed By: dfaure

Subscribers: broulik, ngraham, probono, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D22510