Treat "application/x-ms-dos-executable" as executable on all platforms
Summary:
Because of Wine, *.exe can be executed in almost all desktop platforms.

The dialog will open when a script/exe/binary is clicked.

For scripts, it shows "Open" and "Execute" button. Same as before.

For *.exe on Linux/BSD, it shows only "Execute" button, but it actually does "Open with Wine".

For native binary, it shows only "Execute" button. This is a new difference. I think users should never click and run a binary without warning. Here is already a `showUntrustedProgramWarning` pop up function. But it only popup for binary without execute bit. If you get a binary from *.tar.gz or *.7z, it may already have a execute bit and no popup will be appear.

BUG: 412694

Reviewers: #frameworks, ngraham, dfaure

Reviewed By: ngraham, dfaure

Subscribers: GB_2, ppeter, dfaure, ngraham, broulik, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D24463