Port KSSLD internals from KSslError to QSslError
Summary:
The strings don't match all enum values exactly, as this tries to keep
the same strings for the corresponding enums from KSslError as mapped
in ktcpsocket.cpp.

Mistakes to do this correctly (or the eventual move to QMetaEnum from
QSslError) however are not dangerous, it will merely result in the user
being asked to confirm ignoring the (now more precise) error again, we
will not silently ignore errors we shouldn't be ignoring.

Reviewers: aacid

Reviewed By: aacid

Subscribers: kde-frameworks-devel

Tags: #frameworks

Maniphest Tasks: T11620

Differential Revision: https://phabricator.kde.org/D24931