Ftp ioslave: fix 4 character time interpreted as year
Summary: Previously, the code checked for years by checking for length 4. However this resulted in dates with single digit hours, e.g. 9:37, being interpreted as years. Thus, it would parse the hour as a year, in this case making it year 9. Now it instead checks for years by checking whether there's a colon.

Reviewers: dfaure, #frameworks, sitter, feverfew

Reviewed By: dfaure

Subscribers: ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25966