Adding support for mounting KIOFuse URLs for applications that don't use KIO
Summary:
This patch is required to provide seamless integration with KIOFuse and KIO.

The patch attempts to convert URLs that applications will not understand
natively into local paths based on a KIOFuse mount. If successful, the URL
passed to the application is changed to its location within the KIOFuse mount.
If it is not successful, the URL is not changed.

If KIOFuse is not installed, then kioexec is simply used.

FEATURE: 75324
BUG: 398216
BUG: 330192
CCBUG: 398446
FIXED-IN: 5.66

Test Plan:
Compile KIO with this patch.

Build and install KIOFuse (follow instructions in README): https://invent.kde.org/kde/kio-fuse

Once KIO is built, Dolphin is built with this newly built KIO, and kio-fuse is built,
one should start the kio-fuse process.

The easiest way to test this is via Dolphin (that is built with this patch): `dbus-launch dolphin`.

Once you have done the above one can simply browse to any non-local location in
say, Dolphin, where upon opening a URL in a non-KIO enabled app, it should open
in the KIOFuse mount, except for MTP/Gdrive URLs - those will use KIOExec.
Conversely, KIO URLs in KIO-enabled apps should not use KIOFuse.

Reviewers: fvogt, davidedmundson, dfaure, ngraham

Reviewed By: dfaure, ngraham

Subscribers: alexde, broulik, sitter, davidedmundson, kde-frameworks-devel, ngraham

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D23384