KDirOperator: Use a fixed line height for scroll speed
Summary:
This patch intends to do mostly the same as D19190, but for KDirOperator.
This uses the font's line height multiplied by the amount of scroll lines
from QApplication as a scroll speed for views. This makes the scroll speed
not depend on the icon size, which greatly improves scrolling behaviour.
For more details, see the linked review.

Test Plan:
Run tests/kfilewidgettest_gui and go to a directory with a lot of items.
Scroll around, then change icon size and see that scrolling speed remains
the same.

Reviewers: ngraham, elvisangelaccio

Reviewed By: ngraham, elvisangelaccio

Subscribers: ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25683