Upstream Dolphin's file rename dialog
Summary:
KIO didn't actually have its own rename dialog, which is necessary to implement the requested rename-from-the-file-dialog feature (189482). This patch upstreams Dolphin's dialog so all KIO::FileWidgets users can use it.

The file's contents are copied from Dolphin and updated, but the name is changed from `RenameDialog` to `RenameFileDialog` because `RenameDialog` already exists in KIO (and is actually an //overwrite// dialog, but we can't rename it until KF6).

CCBUG: 189482

Test Plan:
Apply D17596 or D17597 and initiate a rename operation for one or more files. See that it works as expected.

Tests still pass.

Reviewers: #frameworks, #dolphin, broulik, ngraham, dfaure, elvisangelaccio

Reviewed By: #dolphin, ngraham, dfaure, elvisangelaccio

Subscribers: elvisangelaccio, lydia, dfaure, sitter, mitchell, emmanuelp, ltoscano, bruns, meven, dhaumann, pino, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D17595