[KFileFilterCombo] Don't add invalid QMimeType to mimes filter
Summary:
When setting the mime filter, check that the QMimeType we get from
QMimeDatabase::mimeTypeForName() is valid before adding it to the mime
filter list. This fixes a bug where if the QMimeType is invalid its name
(and comment) property is empty, leading to empty entries in the filter
widget drop-down menu.

Per dfaure's recommendation, display a warning about the invalid mimeType.

Also install the necessary Qt logging catergory bits for KFileWidgets.

BUG: 417355

FIXED-IN: 5.68

Test Plan:
- Open gwenview -> file -> open and check the filter drown-down menu
  there are a couple (two?) empty entries
- Apply diff and try again, no empty entries

Reviewers: #frameworks, dfaure, meven, sitter, broulik

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27865