FileWidgets: Ignore Return events from KDirOperator
Summary:
Activated event is used to handle key returns from KDirOperator.
Making a Return in the KDirOperator would cause a double treatment of the event :
 - first KDirOperator::activated for the activated event
 - second KDirOperator::activated for the KeyPressEvent Key_Return

BUG: 412737
FIXED-IN: 5.69

Test Plan:
 - Open Kate
 - Ctrl+Maj + S, to save file as
 - type a file name
 - go to a directory using the arrows keys and return key

Before:
The dialog returns immediately with the file name in the opened directory.

After:
The directory is opened, the user can continue specifying path.

Reviewers: dfaure, ngraham, #frameworks

Reviewed By: dfaure, ngraham

Subscribers: ahmadsamir, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27455