KDirModel: implement showing a root node for the requested URL
Summary:
To trigger this, cal KDirModel::openUrl(rootUrl, KDirModel::ShowRoot)

KDirModel has always been lacking a openUrl method anyway, it's always
been weird to create a KDirModel but then call
dirModel->dirLister()->openUrl().

This feature has been requested by Raphael Rosch who is reimplementing
the konqueror sidebar tree on top of KDirModel. Having a visible item
for "/" is necessary in order to be able to click on it and have it
listed in the main file view.

Test Plan: kdirmodeltest_gui and 2 new methods in kdirmodeltest (unittest)

Reviewers: stefanocrocco, elvisangelaccio, meven, apol

Subscribers: rrosch, ahmadsamir, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D25315