[KOpenWithDialog] Add generic name from .desktop files as a tooltip
Summary:
Extend the current setup by adding a tooltip based on the generic name
entry from the .desktop files.

This is especially useful for new users who don't know what each application
in the list can be used for.

CCBUG: 109016

Test Plan:
make && ctest
And a quick test of the openwith dialog, a tooltip is displayed for
most entries.

Reviewers: #frameworks, dfaure

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28371