KPropertiesDialog: port KRun::run to ApplicationLauncherJob
Summary:
KRun::run(exec, urls) is somewhere in between the two jobs.
Since we know the name of the desktop file, we might as well use ApplicationLauncherJob here.
Otherwise we would probably need a bit of KShell escaping to turn a QUrl into a command-line arg.

Test Plan:
With filelight installed, the button in the properties dialog works.
With the .desktop file still installed but the executable renamed, an error message pops up as expected.

Reviewers: davidedmundson, broulik

Reviewed By: broulik

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28266