autotests: repair OpenUrlJobTest
* runNativeExecutable was always setting a TestUntrustedProgramHandler
even when withHandler was false

* openOrExecuteDesktop didn't make the desktop file executable,
and this worked "by chance" previously because the
TestUntrustedProgramHandler from runNativeExecutable was still
set globally. With the new design where handlers are part of
the job delegate, we'd have to either set a handler again here,
or well, just make the desktop file executable, it's not what this test
is about (it's already testing TestOpenOrExecuteHandler, too many
combinations if we bring a TestUntrustedProgramHandler into the mix).