Mark KIOFuse mounts as Probably slow
Summary:
Mark KIOFuse mounts as "Probably Slow". Note that this isn't always true,
there are plenty of slaves that simply refer to local files that can be
mounted in a KIOFuse mounts, but likewise, there are many slaves that
refer to network mounts that are slow, so better safe than sorry. And in
any case, the function is "probablySlow()", we don't have to confirm that
this is actually the case.

Test Plan:
Compiles.
KIOFuse mounts indeed have a mount type of "fuse.kio-fuse"

Reviewers: #frameworks, dfaure, broulik, bruns, meven

Reviewed By: broulik, meven

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28440