[DirectorySizeJob] Fix sub-dirs count when resolving symlinks to dirs
Summary:
After 730a6ddd828fb1fbdf0f3, the behaviour changed and the code sets
KIO::UDSEntry::UDS_DEVICE_ID after symlinks are resovled (if
KIO::StatResolveSymlink is set); this meant that if we have dir A and
a symlink to it B, the loop that detects hard-links will skip A if B
was processed before, as both will have the same inode.

This fixes the directorySize() unit test from kiocore-jobtest; the
totalSubdirs() count didn't match the expected value.

Test Plan: make && ctest

Reviewers: #frameworks, dfaure, meven

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28472