[KPropertiesDialog] Disable changing remote dir icons
Summary:
Changing the icon of a dir on e.g. a samba share doesn't seem to have
an effect, this is intentional as reading .directory would impact network
resources negatively, see the bug report for more details. Change the code
to only show the KIconButton only for local dirs and desktop:/.

BUG: 205954

Test Plan:
make && ctest
And testing an sftp://, the button to change the icon isn't shown

Reviewers: #frameworks, dfaure, sitter

Reviewed By: dfaure

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D28679