Remove dead code since KF5.0: mount/umount devices in their contextmenu
Summary:
This code was disabled in preparation for KF 5.0 to remove kio's dependency on solid.
This is ironic because later on kio_trash brought that dependency again. But anyway, this
feature was still dead code for all of KF5 life time.

The idea of going to /dev and right-click mount/unmount just seems very weird as a GUI feature.

(Ignore the comment in the CMakeLists.txt, someone misunderstood what that code was about.
No desktop files involved there, this was about block devices themselves)

Test Plan: It builds ;)

Reviewers: apol, broulik

Reviewed By: apol

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29216